+++ gherkin/go/dialects_builtin.go
@@ -1,6 +1,6 @@
 package gherkin
 
-// Builtin dialects for af (Afrikaans), am (Armenian), an (Aragonese), ar (Arabic), ast (Asturian), az (Azerbaijani), bg (Bulgarian), bm (Malay), bs (Bosnian), ca (Catalan), cs (Czech), cy-GB (Welsh), da (Danish), de (German), el (Greek), em (Emoji), en (English), en-Scouse (Scouse), en-au (Australian), en-lol (LOLCAT), en-old (Old English), en-pirate (Pirate), eo (Esperanto), es (Spanish), et (Estonian), fa (Persian), fi (Finnish), fr (French), ga (Irish), gj (Gujarati), gl (Galician), he (Hebrew), hi (Hindi), hr (Croatian), ht (Creole), hu (Hungarian), id (Indonesian), is (Icelandic), it (Italian), ja (Japanese), jv (Javanese), ka (Georgian), kn (Kannada), ko (Korean), lt (Lithuanian), lu (Luxemburgish), lv (Latvian), mk-Cyrl (Macedonian), mk-Latn (Macedonian (Latin)), mn (Mongolian), nl (Dutch), no (Norwegian), pa (Panjabi), pl (Polish), pt (Portuguese), ro (Romanian), ru (Russian), sk (Slovak), sl (Slovenian), sr-Cyrl (Serbian), sr-Latn (Serbian (Latin)), sv (Swedish), ta (Tamil), th (Thai), tl (Telugu), tlh (Klingon), tr (Turkish), tt (Tatar), uk (Ukrainian), ur (Urdu), uz (Uzbek), vi (Vietnamese), zh-CN (Chinese simplified), zh-TW (Chinese traditional)
+// Builtin dialects for af (Afrikaans), am (Armenian), an (Aragonese), ar (Arabic), ast (Asturian), az (Azerbaijani), bg (Bulgarian), bm (Malay), bs (Bosnian), ca (Catalan), cs (Czech), cy-GB (Welsh), da (Danish), de (German), el (Greek), em (Emoji), en (English), en-Scouse (Scouse), en-au (Australian), en-lol (LOLCAT), en-old (Old English), en-pirate (Pirate), eo (Esperanto), es (Spanish), et (Estonian), fa (Persian), fi (Finnish), fr (French), ga (Irish), gj (Gujarati), gl (Galician), he (Hebrew), hi (Hindi), hr (Croatian), ht (Creole), hu (Hungarian), id (Indonesian), is (Icelandic), it (Italian), ja (Japanese), jv (Javanese), ka (Georgian), kn (Kannada), ko (Korean), lt (Lithuanian), lu (Luxemburgish), lv (Latvian), mk-Cyrl (Macedonian), mk-Latn (Macedonian (Latin)), mn (Mongolian), ne (Nepali), nl (Dutch), no (Norwegian), pa (Panjabi), pl (Polish), pt (Portuguese), ro (Romanian), ru (Russian), sk (Slovak), sl (Slovenian), sr-Cyrl (Serbian), sr-Latn (Serbian (Latin)), sv (Swedish), ta (Tamil), th (Thai), tl (Telugu), tlh (Klingon), tr (Turkish), tt (Tatar), uk (Ukrainian), ur (Urdu), uz (Uzbek), vi (Vietnamese), zh-CN (Chinese simplified), zh-TW (Chinese traditional)
 func GherkinDialectsBuildin() GherkinDialectProvider {
 	return buildinDialects
 }
@@ -2354,6 +2354,54 @@ var buildinDialects = gherkinDialectMap{
 				"* ",
 				"Гэхдээ ",
 				"Харин ",
+			},
+		},
+	},
+	"ne": &GherkinDialect{
+		"ne", "Nepali", "नेपाली", map[string][]string{
+			feature: []string{
+				"सुविधा",
+				"विशेषता",
+			},
+			rule: []string{
+				"नियम",
+			},
+			background: []string{
+				"पृष्ठभूमी",
+			},
+			scenario: []string{
+				"परिदृश्य",
+			},
+			scenarioOutline: []string{
+				"परिदृश्य रूपरेखा",
+			},
+			examples: []string{
+				"उदाहरण",
+				"उदाहरणहरु",
+			},
+			given: []string{
+				"* ",
+				"दिइएको ",
+				"दिएको ",
+				"यदि ",
+			},
+			when: []string{
+				"* ",
+				"जब ",
+			},
+			then: []string{
+				"* ",
+				"त्यसपछि ",
+				"अनी ",
+			},
+			and: []string{
+				"* ",
+				"र ",
+				"अनी ",
+			},
+			but: []string{
+				"* ",
+				"तर ",
 			},
 		},
 	},